Added plugin-api module and moved api interfaces to api module
