+++ components/plugin-api/build.gradle
@@ -1,0 +1,14 @@
+dependencies {
+
+}
+
+publishing {
+    publications {
+        mavenJava(MavenPublication) {
+            pom {
+                name = 'Yaci Store - Plugin Api Module'
+                description = 'Yaci Store - Plugin Api Module'
+            }
+        }
+    }
+}
