+++ components/plugin-api/src/main/java/com/bloxbean/cardano/yaci/store/plugin/api/PreActionPlugin.java
