+++ components/plugin-api/src/main/java/com/bloxbean/cardano/yaci/store/plugin/api/VariableProvider.java
@@ -1,4 +1,4 @@
-package com.bloxbean.cardano.yaci.store.plugin.variables;
+package com.bloxbean.cardano.yaci.store.plugin.api;
 
 import java.util.Map;
