+++ components/plugin-polyglot/src/main/java/com/bloxbean/cardano/yaci/store/plugin/polyglot/common/GlobalScriptContextRegistry.java
@@ -1,6 +1,6 @@
 package com.bloxbean.cardano.yaci.store.plugin.polyglot.common;
 
-import com.bloxbean.cardano.yaci.store.common.plugin.ScriptRef;
+import com.bloxbean.cardano.yaci.store.plugin.api.config.ScriptRef;
 import org.graalvm.polyglot.Context;
 
 import java.util.Map;
