+++ components/plugin/src/main/java/com/bloxbean/cardano/yaci/store/plugin/impl/mvel/MvelStorePluginFactory.java
@@ -1,7 +1,7 @@
 package com.bloxbean.cardano.yaci.store.plugin.impl.mvel;
 
-import com.bloxbean.cardano.yaci.store.common.plugin.PluginDef;
-import com.bloxbean.cardano.yaci.store.common.plugin.ScriptRef;
+import com.bloxbean.cardano.yaci.store.plugin.api.config.PluginDef;
+import com.bloxbean.cardano.yaci.store.plugin.api.config.ScriptRef;
 import com.bloxbean.cardano.yaci.store.plugin.api.*;
 import com.bloxbean.cardano.yaci.store.plugin.cache.PluginStateService;
 import com.bloxbean.cardano.yaci.store.plugin.variables.VariableProviderFactory;
