+++ components/plugin/src/main/java/com/bloxbean/cardano/yaci/store/plugin/impl/spel/SpelStorePluginFactory.java
@@ -1,7 +1,7 @@
 package com.bloxbean.cardano.yaci.store.plugin.impl.spel;
 
-import com.bloxbean.cardano.yaci.store.common.plugin.PluginDef;
-import com.bloxbean.cardano.yaci.store.common.plugin.ScriptRef;
+import com.bloxbean.cardano.yaci.store.plugin.api.config.PluginDef;
+import com.bloxbean.cardano.yaci.store.plugin.api.config.ScriptRef;
 import com.bloxbean.cardano.yaci.store.plugin.api.*;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.expression.Expression;
