+++ components/plugin/src/main/java/com/bloxbean/cardano/yaci/store/plugin/variables/DefaultVariableProvider.java
@@ -1,5 +1,6 @@
 package com.bloxbean.cardano.yaci.store.plugin.variables;
 
+import com.bloxbean.cardano.yaci.store.plugin.api.VariableProvider;
 import com.bloxbean.cardano.yaci.store.plugin.cache.PluginStateService;
 import com.bloxbean.cardano.yaci.store.plugin.util.PluginContextUtil;
 import lombok.RequiredArgsConstructor;
