+++ settings.gradle
@@ -24,6 +24,7 @@ include 'components:rocksdb'
 include 'components:job'
 include 'components:admin'
 include 'components:dbutils'
+include 'components:plugin-api'
 include 'components:plugin'
 include 'components:plugin-polyglot'
 include 'components:plugin-model-types'
