+++ stores/metadata/src/main/java/com/bloxbean/cardano/yaci/store/metadata/plugin/MetadataPluginEventHandlers.java
@@ -2,7 +2,7 @@ package com.bloxbean.cardano.yaci.store.metadata.plugin;
 
 import com.bloxbean.cardano.yaci.store.common.config.StoreProperties;
 import com.bloxbean.cardano.yaci.store.metadata.domain.TxMetadataEvent;
-import com.bloxbean.cardano.yaci.store.plugin.api.PluginRegistry;
+import com.bloxbean.cardano.yaci.store.plugin.core.PluginRegistry;
 import com.bloxbean.cardano.yaci.store.plugin.events.PluginBaseEventHandler;
 import org.springframework.context.event.EventListener;
 import org.springframework.stereotype.Component;
