+++ stores/staking/src/test/java/com/bloxbean/cardano/yaci/store/staking/SpringBootTestApplication.java
@@ -2,7 +2,7 @@ package com.bloxbean.cardano.yaci.store.staking;
 
 import com.bloxbean.cardano.yaci.store.common.config.StoreProperties;
 import com.bloxbean.cardano.yaci.store.epoch.service.ProtocolParamService;
-import com.bloxbean.cardano.yaci.store.plugin.api.PluginRegistry;
+import com.bloxbean.cardano.yaci.store.plugin.core.PluginRegistry;
 import com.bloxbean.cardano.yaci.store.staking.service.DepositParamService;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
