+++ stores/utxo/src/main/java/com/bloxbean/cardano/yaci/store/utxo/plugin/UtxoPluginEventHandlers.java
@@ -1,7 +1,7 @@
 package com.bloxbean.cardano.yaci.store.utxo.plugin;
 
 import com.bloxbean.cardano.yaci.store.common.config.StoreProperties;
-import com.bloxbean.cardano.yaci.store.plugin.api.PluginRegistry;
+import com.bloxbean.cardano.yaci.store.plugin.core.PluginRegistry;
 import com.bloxbean.cardano.yaci.store.plugin.events.PluginBaseEventHandler;
 import com.bloxbean.cardano.yaci.store.utxo.domain.AddressUtxoEvent;
 import org.springframework.context.event.EventListener;
